fix: error message was swallow by JSON.stringify
closes #602

after trying to reproduce #602 i found that it already was fixed, but
there was still an issue where the errormessage wasn't found in the
logs. issue was `typeof Error` return `object` so the logger treated is
as object and tried to JSON.stringify it. that made it to display just
`{}`