fix: Error type serialization in logger (#634)
After trying to reproduce #602 i found that it already was fixed, but
there was still an issue where the error message wasn't found in the
logs. issue was `typeof Error` return `object` so the logger treated is
as object and tried to JSON.stringify it. that made it to display just `{}`