style: Prevents "sticky" hover on mobile
Used media queries to remove it from mobile.

Hover effect doesn't work properly on mobile browsers and gets "sticky"
when element (generally buttons) is touched, forcing user to touch
outside of the element to disable its "hover" effect.

Not necessarily a problem everywhere (e.g., should be fine on nav and
lists, so it was kept for such cases).