feat: mnemonic as alternative to download key file (#910)
* feat: Support nested i18n translation objects

* feat: Conditionally remove white spaces from

* test: Use `removeSpaces` in `CopyButton` tests

* style:  Add CSS class to invoice input

Applying a class to narrow down the CSS application, instead of applying
to all `textarea` elements

* feat: Add option to display mnemonic as alternative to download key file

* test: Add mnemonic input/verification unit tests

* style: Adjust spacing in RefundExternal page

* style: Prevents "sticky" hover on mobile

Used media queries to remove it from mobile.

Hover effect doesn't work properly on mobile browsers and gets "sticky"
when element (generally buttons) is touched, forcing user to touch
outside of the element to disable its "hover" effect.

Not necessarily a problem everywhere (e.g., should be fine on nav and
lists, so it was kept for such cases).

* fix: Flaky redirect to home when creating swaps

Due to reactivity race conditions, sometimes `valid()` would be set to
`false` in `Backup.tsx`.

However, it is not necessary to validate the swap again in `backupDone`,
because it is only invoked after the swap was already validated.

* fix: Add missing CSS variables

* chore: Update i18n translations