fix: prevent double read of json stream
When backend returned 502, `json()` was being invoked twice on response:
at `helper.ts` and then `rescue.ts`, which was causing an uncaught
exception because json stream can't be read twice (unless cloned).

The first `.json` invoke was also incorrect because the response was
actually an HTML. Now it validates the content type before parsing it.